requirements.txt
streamlit
pypdf
pandas
google-genai
numpy